Fix to support 2GB HPS DDR
The DMA opcode buffer used for scrubbing the SDRAM was too small to support the
larger RAM size. To fix this the buffer was increased. Also scrubbing
the RAM takes 3 seconds, where the watchdog was set to 2 seconds. The
watchdog is now set to 4 seconds in order to allow for the scrubbing of
the larger RAM to finish.

Signed-off-by: Dan Vincelette <dvincelette@criticallink.com>