MitySOM-5CSx: Factory Configuration Rearchitecture
With the introduction of factoryconfig v1.3, there was a firmware
backwards compatibilty introduced. In order to fix this issue the main
factoryconfig has been changed back to v1.2. Any additional data will be
added to generic configuration blocks but the base factoryconfig will
stay the same. The difference between v1.3 and v1.2 was the addition of
the 2nd MAC address. This will now be stored in a generic configuration
block. Parsing of v1.3 structures will still be supported but will be
written to EEPROM as v1.2 with the 2nd MAC address stored in a generic
configuration block.