FogBugz #470924: U-Boot: Changes for gcc6.2.0
Minor change needed to support the GCC 6.2.0 compiler with
our older socfpga_2013.01.01 U-Boot branch for Cyclone5.
Change is to add a header file that the older branch build expects.
This file is simply a modified copy of the compiler-gcc5.h file
as suggested on the embedded forums. Modified copy to change
the comment at the top from gcc5 to gcc6.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>